requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.in'
#
about-time==4.2.1
    # via alive-progress
alive-progress==3.1.5
    # via -r .\requirements.in
click==8.1.7
    # via typer
colorama==0.4.6
    # via click
grapheme==0.6.0
    # via alive-progress
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
pillow==10.4.0
    # via -r .\requirements.in
pygments==2.18.0
    # via rich
pyyaml==6.0.1
    # via -r .\requirements.in
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
svgwrite==1.4.3
    # via -r .\requirements.in
typer==0.12.3
    # via -r .\requirements.in
typing-extensions==4.12.2
    # via typer
